move csharp extraction rules behind the indexer language registry
S3-CSHARP (ARCH-012), one more one-language rollback unit and a sibling of
S3-KOTLIN. C#'s tree-sitter rule file, compiled-rule cache, member-call callsite
marker, and receiver-call resolution engine move out of the indexer god file
into `codestory-indexer/src/languages/csharp.rs`, behind the
`LanguageExtraction` registry in `languages/mod.rs`. Every language-keyed
dispatch that used to spell "csharp" now reads the registry first and keeps its
residual `match` for the languages that have not moved; the last sibling package
deletes those arms.

The move is mechanical: 712 non-blank lines left `lib.rs` (28317 -> 27575 lines,
26613 -> 25901 non-blank). 699 of them are the function bodies now in
`languages/csharp.rs`, and `diff` between the deleted lines and the moved bodies
is empty once the two declared renames are undone
(`collect_csharp_receiver_call_edges` -> `pub(crate) receiver_call_specs`,
`csharp_member_call` -> `member_call`, two call sites). Twenty-six functions and
two structs left `lib.rs` and the same twenty-six functions and two structs
landed in the module, plus the new `csharp_language()` parser constructor. The
remaining 13 non-blank lines are the residual roster rows: the marker const, the
graph-query const, the rules cache static, the receiver-call arm, the
`csharp_member` call-syntax arm, the reflowed `route_language_uses_c_style_comments`
list, and the `LanguageRuleset::CSharp` arm turned fail-closed.

Proof: `tests/language_extraction_snapshot.rs` gains one `SnapshotCase` row that
pins the full rendered projection of both C# fixtures -- the fidelity_lab file
through the store pipeline (resolution included) and the tictactoe file through
raw `index_file`. Both goldens were captured before the production edit and are
byte-identical after it (sha256 6b5b4ca2.. / 1e41a32c..). The threshold suites
are not sufficient on their own: changing C#'s qualified-name delimiter to "::"
leaves fidelity_regression (8 passed), tictactoe_language_coverage (12 passed)
and all 269 lib unit tests green while both new snapshots go red on
`qn=TicTacToe::Entry` versus `qn=TicTacToe.Entry`.

Declared behavior changes:
- None to any indexing, resolution, or rendering output. C#
  fidelity_regression and tictactoe_language_coverage projections are proven
  byte-identical before and after the move, and packet_search_eval is unmoved.
- `LanguageRuleset::CSharp`'s arm in `compiled_rules` is now unreachable and
  returns a typed error instead of compiled rules; the registry answers first.
  Fail-closed by construction, exactly as the Kotlin arm already is.
- One additive `LanguageCommentProfile` row for "csharp" in
  `codestory_contracts::language_support`, and the matching deletion of "csharp"
  from the CLI's residual comment roster in `output.rs::ansi_highlight_line`.
  The marker is "//" on both sides, so rendering is unchanged; the restated
  pre-move roster in `comment_markers_stay_identical_after_the_registry_lookup`
  still passes unedited. The CLI roster and the indexer route-comment roster
  still disagree on cpp/swift/tsx/jsx and vue/astro; those are separate facts
  with separate owners and are left alone.
- `LANGUAGE_SUPPORT_PROFILES` is unchanged, so extension routing, claim tiers,
  `supported_extensions`, `is_structural_source_path`, and every existing
  language claim are identical. The row claims only `cs`; `cshtml` stays a
  companion Razor extension with no parser-backed C# claim.
- `semantic::CSharpSemanticResolver` stays behind `dedicated_semantic_resolver`
  because the resolver types are private to that module; the registry records
  the choice as `uses_generic_semantic_resolver: false`.
- No version surface touched; everything stays 0.16.3.

Closes #1688
Refs #1243
Refs #1179
