+++ crates/codestory-indexer/src/semantic/mod.rs
@@ -599,7 +599,6 @@ fn language_family_bucket(language: &'static str) -> &'static str {
         "go" => "go",
         "ruby" => "ruby",
         "php" => "php",
-        "csharp" => "csharp",
         "swift" => "swift",
         "dart" => "dart",
         "bash" => "bash",
